recommender/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
surprise==0.1
numpy>=1.26.0
scipy==1.11.4
python-dotenv==1.0.0
pandas==2.1.4
scikit-learn==1.3.2